move transformers.ts
