+++ src/lib/SideBars/CharConfig.svelte
@@ -29,7 +29,7 @@
     import TriggerList from "./Scripts/TriggerList.svelte";
   import CheckInput from "../UI/GUI/CheckInput.svelte";
   import { updateInlayScreen } from "src/ts/process/inlayScreen";
-  import { registerOnnxModel } from "src/ts/process/embedding/transformers";
+  import { registerOnnxModel } from "src/ts/process/transformers";
   import MultiLangInput from "../UI/GUI/MultiLangInput.svelte";
